add eslint rules and fix lint errors
