+++ .eslintrc.json
@@ -1,0 +1,24 @@
+{
+    "extends": [
+      "airbnb",
+      "airbnb/hooks",
+      "eslint:recommended",
+      "react-app"
+    ],
+    "env": {
+      "browser": true,
+      "commonjs": true,
+      "node": true,
+      "es6": true
+    },
+    "parserOptions": {
+      "ecmaVersion": 2018,
+      "sourceType": "module"
+    },
+    "rules": {
+      "curly": "warn",
+	  "react/jsx-indent": [2, "tab"],
+	  "no-tabs": "off",
+	  "max-len": "off"
+    }
+  }
